Clean up tests and CI

+++ packages/matter.js/src/util/ByteArray.ts
@@ -29,7 +29,7 @@ declare global {
 }
 
 Uint8Array.prototype.toHex = function() {
-    return this.reduce((result, byte) => result + byte.toString(16).padStart(2, "0"), "");
+    return (this as Uint8Array).reduce((result, byte) => result + byte.toString(16).padStart(2, "0"), "");
 }
 
 Uint8Array.prototype.getDataView = function() {
@@ -38,7 +38,7 @@ Uint8Array.prototype.getDataView = function() {
 
 Uint8Array.prototype.equals = function(other: Uint8Array) {
     if (other.length !== this.length) return false;
-    return this.every((value, index) => other[index] === value);
+    return (this as Uint8Array).every((value, index) => other[index] === value);
 }
 
 Uint8Array.fromHex = function(hexString: string) {
